Started pulling season methods off stat_tracker
