+++ lib/stat_tracker.rb
@@ -208,119 +208,119 @@ class StatTracker
 
   ### Begin team stat methods
 
-  def team_info(team_id)
-    x = {}
-    teams.find_all do |team|
-      if team.team_id == team_id.to_s
-        x = team.instance_variables.each_with_object({}) { |var, hash| hash[var.to_s.delete("@")] = team.instance_variable_get(var) }
-        return x
-      end
-    end
-  end
-
-  def find_number_of_games_played_in_a_season(team_id)
-    games_played_that_season = Hash.new(0)
-    game_teams.find_all do |team|
-      if team.team_id == team_id.to_s
-          games_played_that_season[team.game_id.slice(0...4)] += 1.to_f
-      end
-    end
-    games_played_that_season
-  end
-
-  def best_season(team_id)
-    games_won_that_season = Hash.new(0)
-    game_teams.find_all do |team|
-      if team.team_id == team_id.to_s
-        if team.result == "WIN"
-          games_won_that_season[team.game_id.slice(0...4)] += 1.to_f
-        end
-      end
-    end
-    games_won_that_season.merge!(find_number_of_games_played_in_a_season(team_id)) { |k, o, n| o / n }
-    ((games_won_that_season.key(games_won_that_season.values.max) * 2).to_i + 1).to_s
-  end
-
-  def worst_season(team_id)
-    games_lost_that_season = Hash.new(0)
-    game_teams.find_all do |team|
-      if team.team_id == team_id.to_s
-        if team.result == "LOSS"
-          games_lost_that_season[team.game_id.slice(0...4)] += 1.to_f
-        end
-      end
-    end
-    games_lost_that_season.merge!(find_number_of_games_played_in_a_season(team_id)) { |k, o, n| o / n }
-    ((games_lost_that_season.key(games_lost_that_season.values.min) * 2).to_i + 1).to_s
-  end
-
-  def average_win_percentage(team_id)
-    games_won_that_season = Hash.new(0)
-    game_teams.find_all do |team|
-      if team.team_id == team_id.to_s
-        if team.result == "WIN"
-          games_won_that_season[team.game_id.slice(0...4)] += 1.to_f
-        end
-      end
-    end
-    (games_won_that_season.values.sum / find_number_of_games_played_in_a_season(team_id).values.sum * 100).round(2)
-  end
-
-  def goals_scored(team_id)
-    all_goals_scored = []
-      game_teams.map do |team|
-        if team.team_id == team_id.to_s
-        all_goals_scored << team.goals.to_i
-      end
-    end
-    all_goals_scored
-  end
-
-  def most_goals_scored(team_id)
-    goals_scored(team_id).max
-  end
-
-  def fewest_goals_scored(team_id)
-    goals_scored(team_id).min
-  end
-
-  def favorite_opponent(team_id)
-    games_won_vs_oppenent = Hash.new(0)
-    games.map do |team|
-      if team.home_team_id || team.away_team_id == team_id.to_s
-        if (team.home_team_id == team_id.to_s) && (team.home_goals > team.away_goals)
-          games_won_vs_oppenent[team.away_team_id] += 1
-        else (team.away_team_id == team_id.to_s) && (team.home_goals < team.away_goals)
-          games_won_vs_oppenent[team.home_team_id] += 1
-        end
-      end
-    end
-    id = games_won_vs_oppenent.key(games_won_vs_oppenent.values.max)
-    found = teams.find do |team|
-      if team.team_id == id
-        return team.team_name
-      end
-      found
-    end
-  end
-
-  def rival(team_id)
-    games_lost_vs_oppenent = Hash.new(0)
-    games.map do |team|
-      if team.home_team_id || team.away_team_id == team_id.to_s
-        if (team.home_team_id == team_id.to_s) && (team.home_goals < team.away_goals)
-          games_lost_vs_oppenent[team.away_team_id] += 1
-        else (team.away_team_id == team_id.to_s) && (team.home_goals > team.away_goals)
-          games_lost_vs_oppenent[team.home_team_id] += 1
-        end
-      end
-    end
-    id = games_lost_vs_oppenent.key(games_lost_vs_oppenent.values.max)
-    found = teams.find do |team|
-      if team.team_id == id
-        return team.team_name
-      end
-      found
-    end
-  end
+  # def team_info(team_id)
+  #   x = {}
+  #   teams.find_all do |team|
+  #     if team.team_id == team_id.to_s
+  #       x = team.instance_variables.each_with_object({}) { |var, hash| hash[var.to_s.delete("@")] = team.instance_variable_get(var) }
+  #       return x
+  #     end
+  #   end
+  # end
+  #
+  # def find_number_of_games_played_in_a_season(team_id)
+  #   games_played_that_season = Hash.new(0)
+  #   game_teams.find_all do |team|
+  #     if team.team_id == team_id.to_s
+  #         games_played_that_season[team.game_id.slice(0...4)] += 1.to_f
+  #     end
+  #   end
+  #   games_played_that_season
+  # end
+  #
+  # def best_season(team_id)
+  #   games_won_that_season = Hash.new(0)
+  #   game_teams.find_all do |team|
+  #     if team.team_id == team_id.to_s
+  #       if team.result == "WIN"
+  #         games_won_that_season[team.game_id.slice(0...4)] += 1.to_f
+  #       end
+  #     end
+  #   end
+  #   games_won_that_season.merge!(find_number_of_games_played_in_a_season(team_id)) { |k, o, n| o / n }
+  #   ((games_won_that_season.key(games_won_that_season.values.max) * 2).to_i + 1).to_s
+  # end
+  #
+  # def worst_season(team_id)
+  #   games_lost_that_season = Hash.new(0)
+  #   game_teams.find_all do |team|
+  #     if team.team_id == team_id.to_s
+  #       if team.result == "LOSS"
+  #         games_lost_that_season[team.game_id.slice(0...4)] += 1.to_f
+  #       end
+  #     end
+  #   end
+  #   games_lost_that_season.merge!(find_number_of_games_played_in_a_season(team_id)) { |k, o, n| o / n }
+  #   ((games_lost_that_season.key(games_lost_that_season.values.min) * 2).to_i + 1).to_s
+  # end
+  #
+  # def average_win_percentage(team_id)
+  #   games_won_that_season = Hash.new(0)
+  #   game_teams.find_all do |team|
+  #     if team.team_id == team_id.to_s
+  #       if team.result == "WIN"
+  #         games_won_that_season[team.game_id.slice(0...4)] += 1.to_f
+  #       end
+  #     end
+  #   end
+  #   (games_won_that_season.values.sum / find_number_of_games_played_in_a_season(team_id).values.sum * 100).round(2)
+  # end
+  #
+  # def goals_scored(team_id)
+  #   all_goals_scored = []
+  #     game_teams.map do |team|
+  #       if team.team_id == team_id.to_s
+  #       all_goals_scored << team.goals.to_i
+  #     end
+  #   end
+  #   all_goals_scored
+  # end
+  #
+  # def most_goals_scored(team_id)
+  #   goals_scored(team_id).max
+  # end
+  #
+  # def fewest_goals_scored(team_id)
+  #   goals_scored(team_id).min
+  # end
+  #
+  # def favorite_opponent(team_id)
+  #   games_won_vs_oppenent = Hash.new(0)
+  #   games.map do |team|
+  #     if team.home_team_id || team.away_team_id == team_id.to_s
+  #       if (team.home_team_id == team_id.to_s) && (team.home_goals > team.away_goals)
+  #         games_won_vs_oppenent[team.away_team_id] += 1
+  #       else (team.away_team_id == team_id.to_s) && (team.home_goals < team.away_goals)
+  #         games_won_vs_oppenent[team.home_team_id] += 1
+  #       end
+  #     end
+  #   end
+  #   id = games_won_vs_oppenent.key(games_won_vs_oppenent.values.max)
+  #   found = teams.find do |team|
+  #     if team.team_id == id
+  #       return team.team_name
+  #     end
+  #     found
+  #   end
+  # end
+  #
+  # def rival(team_id)
+  #   games_lost_vs_oppenent = Hash.new(0)
+  #   games.map do |team|
+  #     if team.home_team_id || team.away_team_id == team_id.to_s
+  #       if (team.home_team_id == team_id.to_s) && (team.home_goals < team.away_goals)
+  #         games_lost_vs_oppenent[team.away_team_id] += 1
+  #       else (team.away_team_id == team_id.to_s) && (team.home_goals > team.away_goals)
+  #         games_lost_vs_oppenent[team.home_team_id] += 1
+  #       end
+  #     end
+  #   end
+  #   id = games_lost_vs_oppenent.key(games_lost_vs_oppenent.values.max)
+  #   found = teams.find do |team|
+  #     if team.team_id == id
+  #       return team.team_name
+  #     end
+  #     found
+  #   end
+  # end
 end
